requirements.txt
libvirt-python==4.3.0
lxml==4.2.1
virtualenv==15.2.0
